requirements.txt
django>=1.11.17,<2.0
Pillow
psycopg2==2.7.3.2
ipython==6.2.1
python-memcached==1.59
xlwt
django-solo==1.1.3
django-render-block==0.5
easy-thumbnails==2.6
django-import-export
wordcloud
git+https://github.com/mpcabd/python-arabic-reshaper
git+https://github.com/MeirKriheli/python-bidi